Reformat source code using clang-format
At the end of the day I kind of hate this. It barely works as a style checker, let alone an actual formatting tool for anything serious.
I intended to use something that would auto-format all the code, but this needs so many exceptions I will not use it much after this initial reformat.